+++ src/pages/Registration/sections/RegistrationInfo.js
