+++ src/pages/Registration/utils.js
@@ -1,0 +1,95 @@
+import { checkEmailPattern, checkMobilePattern } from "src/utils/inputValidator";
+
+const required = 'يجب تعبئة الحقل'
+
+const CitizenValidate = values => {
+  var msg = {}
+  if (!values.idNumber)
+    msg.idNumber = required;
+  console.log(`nationality ${JSON.stringify(values.nationality)}`);
+  if (values.nationality === "SA") {
+    if (!values.day)
+      msg.day = required;
+    if (!values.month)
+      msg.month = required;
+    if (!values.year)
+      msg.year = required;
+  }
+  console.log(`mssssss ${JSON.stringify(msg)}`);
+  return msg
+}
+const absherValidate = values => {
+  var msg = {}
+  if (!values.AbsherOtp)
+    msg.AbsherOtp = required;
+  return msg
+}
+const regitrationValidate = values => {
+  var msg = {}
+  const format = /[ `!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?~]/;
+
+  if (!values.email)
+    msg.email = required;
+  else if (!checkEmailPattern(values.email)){
+    msg.email = 'يرجى ادخال البريد الكتروني صحيح';
+  }
+  if (!values.password)
+    msg.password = required;
+  if (!values.passwordConfirmation)
+    msg.passwordConfirmation = required;
+  if (values.passwordConfirmation && values.passwordConfirmation !== values.password) {
+    console.log(`${values.passwordConfirmation} ${values.password}`);
+    msg.passwordConfirmation = 'كلمة المرور لا تماثل التأكيد';
+  }
+  document.getElementsByTagName('UL')[0].style.color = 'red';
+  if (values.password) {
+    if (values.password.length >= 8)
+      document.getElementById('digitsNo').style.color = '#04AA6D';
+    else {
+      document.getElementById('digitsNo').style.color = 'red';
+      msg.password = 'حقل كلمة المرور غير صحيح';
+    }
+    if (/\d/.test(values.password))
+      document.getElementById('digitExist').style.color = '#04AA6D';
+    else {
+      document.getElementById('digitExist').style.color = 'red';
+      msg.password = 'حقل كلمة المرور غير صحيح';
+    }
+    if (values.password.toUpperCase() !== values.password)
+      document.getElementById('UpperCase').style.color = '#04AA6D';
+    else {
+      document.getElementById('UpperCase').style.color = 'red';
+      msg.password = 'حقل كلمة المرور غير صحيح';
+    }
+    if (values.password.toLowerCase() !== values.password)
+      document.getElementById('LowerCase').style.color = '#04AA6D';
+    else {
+      document.getElementById('LowerCase').style.color = 'red';
+      msg.password = 'حقل كلمة المرور غير صحيح';
+    }
+    if (format.test(values.password))
+      document.getElementById('symbol').style.color = '#04AA6D';
+    else {
+      document.getElementById('symbol').style.color = 'red';
+      msg.password = 'حقل كلمة المرور غير صحيح';
+    }
+  } else {
+    msg.password = required;
+  }
+  console.log('msg', msg)
+  return msg
+}
+
+const TaheelOtpValidate = (values) => {
+  var msg = {}
+  if (!values.phoneNumber)
+    msg.phoneNumber = required;
+  if (!checkMobilePattern(values.phoneNumber)) {
+    msg.phoneNumber = 'يرجى ادخال رقم جوال سعودي صحيح';
+  }
+  if (values.isTaheelValidate && !values.taheelOtp)
+    msg.taheelOtp = required;
+  return msg
+}
+
+export { CitizenValidate, absherValidate, regitrationValidate, TaheelOtpValidate }
